# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Doc",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/mit-license.php"
        },
        "version": "0.0.1"
    },
    "host": "localhost:6969",
    "basePath": "/api/v1",
    "paths": {
        "/employees": {
            "get": {
                "description": "Get all employees",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employees"
                ],
                "summary": "Get all employees",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/employees/{id}": {
            "get": {
                "description": "Get employee by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employees"
                ],
                "summary": "Get employee by id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check status of app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Get status of app",
                "responses": {
                    "200": {
                        "description": "healthy: This application is started.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "None: This application is stopped.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}